% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_fun.R
\name{are_fun}
\alias{are_fun}
\title{Absolute Reconstruction Error for Z, Ps, and Pb}
\usage{
are_fun(true.tree, inferred.tree)
}
\arguments{
\item{true.tree}{an object of class \code{phylo} representing the ground truth
phylogenetic tree}

\item{inferred.tree}{an object of class \code{phylo} representing the inferred
phylogenetic tree}
}
\value{
A list containing: \code{are.Z}, the absolute reconstruction error associated
with inference of the clonal configuration matrix, Z, \code{are.Ps}, the
absolute reconstruction error associated with inference of the cell-to-clone
assignment matrix, Ps, and \code{are.Pb}, the absolute reconstruction error
associated with inference of the sample-to-clone assignment matrix, Pb.
}
\description{
Compute the absolute reconstruction error (ARE) associated with inference of
the clonal configuration matrix, Z, cell-to-clone assignment matrix, Ps,
and sample-to-clone assignment matrix, Pb.
}
\details{
Please refer to Algorithm S3 in the manuscript text for details regarding
how the absolute reconstruction error is computed. In brief, the error for
each component is quantified by finding the minimum absolute distance between
the inference and all permutations of the truth by 1) computing the distance
metric for each component, 2) computing extra error from additional columns,
and 3) estimating the ARE for each component as a combination of the error
from steps 1) and 2).
}
\examples{
# Simulate alternative and total read counts for bulk and single-cell data
sims.out<-simulate_data(N=15, S=2, M=5, alpha=0.1, beta=1, kappa=1, tau=999,
                        Ktrue=4, b.mindepth=30, b.maxdepth=50, sc.mindepth=80,
                        sc.maxdepth=120, scale=300, seed=8675309)

# Estimate parameters for gene kinetics using the BPSC methodology
# Note: can also use get_burstiness_scale() for large datasets, but estimates
# are not as reliable
param.out<-get_burstiness_bpsc(counts=sims.out$G)

# Subset the read counts to include only those mutations with estimable gene
# kinetics (all are estimable here, but included for utility when using real
# data)
Rs<-sims.out$Rs[param.out$id.g,] # single cell alternative read counts
Xs<-sims.out$Xs[param.out$id.g,] # single cell total read counts
Rb<-sims.out$Rb[param.out$id.g,] # bulk alternative read counts
Xb<-sims.out$Xb[param.out$id.g,] # bulk total read counts

# Run Canopy2 to get list of phylogenetic trees corresponding to all chains
# and all subclones
get.trees.out<-get_trees(Rs=Rs, Rb=Rb, Xs=Xs, Xb=Xb,
                        alpha=param.out$alpha, beta=param.out$beta,
                        kappa=1, tau=999, Klist=3:5,
                        niter=5000, nchains=3, thin=10, pburn=0.1,
                        seed=8675309)

# Get best tree across all chains and subclones via DIC
best.tree.out<-get_best_tree(get.trees.out)

# Compute absolute reconstruction error for each component
are.out<-are_fun(true.tree=sims.out$true.tree, inferred.tree=best.tree.out$tree)

# Print absolute reconstruction error for each component
are.out$are.Z; are.out$are.Ps; are.out$are.Pb

}
