% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_data.R
\docType{data}
\name{BC03}
\alias{BC03}
\title{Breast cancer (BC) Data from Patient BC03}
\format{
Labels:
\itemize{
\item Bulk samples (RNA-seq): SRR2973275 (primary sample), SRR2973276 (LN sample)
\item Single-cell samples (RNA-seq): SRR2973351 to SRR2973436 and SRR5023442
\item Blood (normal) samples: SRR3023076
\item Bulk samples (DNA WES): SRR3023078 (LN sample), SRR3023080 (primary sample)
}

\code{BC03}   An object of class \code{S4}. Access internal objects
from slots using @.

\code{BC03@alt.qc}   A matrix with 80 columns and 145 rows. The column
names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood samples
according to the above labels from the SNV callset of patient BC03. The row
names denote the chromosomal position of the point mutations.

\code{BC03@annovar}   A dataframe with 39 columns and 145 rows.
For details regarding the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}.

\code{BC03@featurecounts.qc}   A matrix with 75 columns and 63677 rows.
The column names denote the scRNA-seq samples from the gene expression data
for patient BC03. The row names denote the gene names.

\code{BC03@mut}   A data frame with five variables: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coinicides with
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{BC03@ref.qc}   A matrix with 80 columns and 145 rows. The
column names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood
samples from the total read counts of patient BC03. The row names denote the
chromosomal position of the point mutations.

\code{BC03@summary.qc}   A matrix with 75 columns and 14 rows. The
column names denote the scRNA-seq samples from the gene expression data for
patient BC03. The row names denote the assignment of the alignments.
}
\usage{
BC03
}
\description{
QC’d single-cell RNA sequencing (scRNA-seq), bulk RNA sequencing (bulk RNA-seq),
and bulk DNA whole exome sequencing data (bulk DNA WES) from one patient
(BC03) of 11 individuals (BC01-BC11) with primary breast cancer
\insertCite{Chung2017}{Canopy2}. The single-nucleotide variant (SNV) pipeline
utilized to QC the data is described in Figure S20 of
\insertCite{Wang2021}{Canopy2}. In the main study, the authors sequenced 549
primary breast cancer cells and matched bulk tumors and/or pooled cells from
11 patients, including two lymph node metastases (BC03LN, BC07LN). Their
final dataset after filtering included 515 single cells and 14 bulk samples.
Accession Number: GSE75688
}
\details{
\code{BC03@alt.qc}   QC'd alternative read counts from bulk and
single-cell data for BC03.

\code{BC03@annovar}   Functional annotation of somatic variants for BC03.

\code{BC03@featurecounts.qc}   Mapped reads from single-cell gene
expression data for BC03.

\code{BC03@mut}   Chromosomal positions of point mutations for BC03.

\code{BC03@ref.qc}   QC'd reference read counts from bulk and single-cell
data for BC03.

\code{BC03@summary.qc}   \code{featureCounts} output summary that reports
assignment of alignments to genomic features

\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (QC)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   BC03 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab Metastatic lymph node \tab 1 \tab 1 (pooled bulk RNA) \tab 55 (45) \cr
    \tab Primary tumor \tab 1 \tab 1 (pooled bulk RNA) \tab 37 (30) \cr
    \tab (total) \tab 3 \tab 2 \tab 92 (75) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
