% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_data.R
\docType{data}
\name{BC03_postproc}
\alias{BC03_postproc}
\title{Post-processed breast cancer (BC) data from patient BC03}
\format{
Labels:

\code{BC03_postproc}   An object of class \code{S4}. Access internal
objects from slots using @.

\code{BC03_postproc@annovar.qc}   A 33 x 39 dataframe containing the
post-processed functional annotations of the somatic variants. For details
regarding the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}. Only those
annotations corresponding to mutations with at least one non-zero count
across all cells were retained.

\code{BC03_postproc@featurecounts.qc}   A 63677 x 77 matrix of
post-processed mapped reads from the single-cell gene expression data.
The row names denote the gene names. The column names from the third column
onward denote the scRNA-seq samples from the gene expression data for patient
BC03. Only those cells that have at least one non-zero count across genes
were retained.

\code{BC03_postproc@param.est}   A list of 5 elements: \code{alpha},
\code{beta}, \code{scale}, \code{id.g}, and \code{pct.estimable}. The first
three are of length equal to the number of mutations where \code{alpha} denotes the
mutation-specific gene activation rates, \code{beta} denotes the mutation-specific
gene deactivation rates, and \code{scale} denotes the mutation-specific
transcription rates. The ids in \code{id.g} denote the positions corresponding
to the mutations that were estimable (returning non-NA or non-negative values),
and \code{pct.estimable} denotes the percentage that were estimable.

\code{BC03_postproc@Rb}   A 31 x 2 matrix of post-processed alternative
read counts from the bulk DNA WES data. The row names denote the chromosomal
position of the point mutations. The column names denote the bulk samples
according to the above labels from the SNV callset of patient BC03.
Post-processing involved removing mutations (rows) with zero alternative
read counts across all cells from the single-cell data.

\code{BC03_postproc@Rs}   A 31 x 75 matrix of post-processed alternative
read counts from the scRNA-seq data. The row names denote the chromosomal
position of the point mutations. The column names denote the scRNA-seq samples
according to the above labels from the SNV callset of patient BC03.
Post-processing involved i) removing mutations (rows) with zero alternative
read counts across all cells and ii) removing cells (columns) with zero total
(benign + mutated) read counts across all mutations.

\code{BC03_postproc@Xb}   A 31 x 2 matrix of post-processed total read
counts (benign + mutated) from the bulk DNA WES data. The row names denote
the chromosomal position of the point mutations. The column names denote the
bulk samples according to the above labels from the SNV callset of patient
BC03. Post-processing involved removing mutations (rows) with zero
alternative read counts across all cells from the single-cell data.

\code{BC03_postproc@Xs}   A 31 x 75 matrix of post-processed total read
counts (benign + mutated) from the scRNA-seq data. The row names denote the
chromosomal position of the point mutations. The column names denote the
scRNA-seq samples according to the above labels from the SNV callset of
patient BC03. Post-processing involved i) removing mutations (rows) with zero
alternative read counts across all cells and ii) removing cells (columns)
with zero total (benign + mutated) read counts across all mutations.
}
\usage{
BC03_postproc
}
\description{
Some further data munging of BC03_preproc to get BC03_postproc, which contains
the finalized read counts for Canopy2 functions and vignette.  \if{html}{\out{<br />}}
Accession Number: GSE75688
}
\details{
\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (post-processed)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   BC03 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab Metastatic lymph node \tab 1 \tab 1 (pooled bulk RNA) \tab 55 (45) \cr
    \tab Primary tumor \tab 1 \tab 1 (pooled bulk RNA) \tab 37 (30) \cr
    \tab (total) \tab 3 \tab 2 \tab 92 (75) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
