% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_data.R
\docType{data}
\name{BC03_preproc}
\alias{BC03_preproc}
\title{Pre-processed breast cancer (BC) data from patient BC03}
\format{
Labels:

\code{BC03_preproc}   An object of class \code{S4}. Access internal
objects from slots using @.

\code{BC03_preproc@alt.qc}   A 145 x 80 matrix of pre-processed alternative
read counts from the bulk and single-cell data for BC03. The row names denote
the chromosomal position of the point mutations. The column names denote the
bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood samples according to the
above labels from the SNV callset of patient BC03.

\code{BC03_preproc@annovar}   A 145 x 39 dataframe containing the
functional annotation of the somatic variants for BC03. For details regarding
the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{canopy2}.

\code{BC03_preproc@featurecounts.qc}   A 63677 x 75 matrix of
pre-processed mapped reads from single-cell gene expression data. The row
names denote the gene names. The column names from the third column onward
denote the scRNA-seq samples from the gene expression data for patient BC03.

\code{BC03_preproc@mut}   A 145 x 5 dataframe containing the chromosomal
positions of point mutations for BC03. The columns consist of: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coincides with
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{BC03_preproc@ref.qc}   A 145 x 80 matrix of pre-processed
reference read counts from the bulk and single-cell data for BC03. The
row names denote the chromosomal position of the point mutations. The
column names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood
samples from the total read counts of patient BC03.

\code{BC03_preproc@summary.qc}   A 14 x 75 matrix of the
\code{featureCounts} output summary that reports assignment of alignments
to genomic features. The row names denote the assignment of the alignments.
The column names denote the scRNA-seq samples from the gene expression data
for patient BC03.
}
\usage{
BC03_preproc
}
\description{
Pre-processed single-cell RNA sequencing (scRNA-seq), bulk RNA sequencing
(bulk RNA-seq), and bulk DNA whole exome sequencing data (bulk DNA WES) from
one patient (BC03) of 11 individuals (BC01-BC11) with primary breast cancer
\insertCite{Chung2017}{canopy2}. The single-nucleotide variant (SNV) pipeline
utilized to pre-process the data is described in the main text and supplement.
In the main study, the authors sequenced 549 primary breast cancer cells and
matched bulk tumors and/or pooled cells from 11 patients, including two lymph node
metastases (BC03LN, BC07LN). Their final dataset after filtering included 515
single cells and 14 bulk samples. \if{html}{\out{<br />}}
Accession Number: GSE75688
}
\details{
\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (pre-processed)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   BC03 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab Metastatic lymph node \tab 1 \tab 1 (pooled bulk RNA) \tab 55 (45) \cr
    \tab Primary tumor \tab 1 \tab 1 (pooled bulk RNA) \tab 37 (30) \cr
    \tab (total) \tab 3 \tab 2 \tab 92 (75) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
