% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_data.R
\docType{data}
\name{BC07}
\alias{BC07}
\title{Breast cancer (BC) Data from Patient BC07}
\format{
Labels:
\itemize{
\item Bulk samples (RNA-seq): SRR2973277 (primary sample), SRR2973278 (LN sample)
\item Single-cell samples (RNA-seq): SRR2973437 to SRR2973535 and SRR5023558-SRR5023562
\item Blood (normal) samples: SRR3023081
\item Bulk samples (DNA WES): SRR3023082 (LN sample), SRR3023083 (primary sample)
}

\code{BC07}   An object of class \code{S4}. Access internal objects
from slots using @.

\code{BC07@alt.qc}   A matrix with 106 columns and 79 rows. The column
names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood samples
according to the above labels from the SNV callset of patient BC07. The row
names denote the chromosomal position of the point mutations.

\code{BC07@annovar}   A dataframe with 39 columns and 79 rows.
For details regarding the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}.

\code{BC07@featurecounts.qc}   A matrix with 101 columns and 63677 rows.
The column names denote the scRNA-seq samples from the gene expression data
for patient BC07. The row names denote the gene names.

\code{BC07@mut}   A data frame with five variables: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coinicides with
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{BC07@ref.qc}   A matrix with 106 columns and 79 rows. The
column names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood
samples from the total read counts of patient BC07. The row names denote the
chromosomal position of the point mutations.

\code{BC07@summary.qc}   A matrix with 101 columns and 14 rows. The
column names denote the scRNA-seq samples from the gene expression data for
patient BC07. The row names denote the assignment of the alignments.
}
\usage{
BC07
}
\description{
QC’d single-cell RNA sequencing (scRNA-seq), bulk RNA sequencing (bulk RNA-seq),
and bulk DNA whole exome sequencing data (bulk DNA WES) from one patient
(BC07) of 11 individuals (BC01-BC11) with primary breast cancer
\insertCite{Chung2017}{Canopy2}. The single-nucleotide variant (SNV) pipeline
utilized to QC the data is described in Figure S20 of
\insertCite{Wang2021}{Canopy2}. In the main study, the authors sequenced 549
primary breast cancer cells and matched bulk tumors and/or pooled cells from
11 patients, including two lymph node metastases (BC03LN, BC07LN). Their
final dataset after filtering included 515 single cells and 14 bulk samples.
GEO Accession Number: GSE75688
}
\details{
\code{BC07@alt.qc}   QC'd alternative read counts from bulk and
single-cell data for BC07.

\code{BC07@annovar}   Functional annotation of somatic variants for BC07.

\code{BC07@featurecounts.qc}   Mapped reads from single-cell gene
expression data for BC07.

\code{BC07@mut}   Chromosomal positions of point mutations for BC07.

\code{BC07@ref.qc}   QC'd reference read counts from bulk and single-cell
data for BC07.

\code{BC07@summary.qc}   \code{featureCounts} output summary that reports
assignment of alignments to genomic features

\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (QC)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   BC07 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab Metastatic lymph node \tab 1 \tab 1 (pooled bulk RNA) \tab 53 (52) \cr
    \tab Primary tumor \tab 1 \tab 1 (tumor bulk RNA) \tab 51 (49) \cr
    \tab (total) \tab 3 \tab 2 \tab 104 (101) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
