% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_data.R
\docType{data}
\name{BC07_preproc}
\alias{BC07_preproc}
\title{Pre-processed breast cancer (BC) Data from Patient BC07}
\format{
Labels:

\code{BC07_preproc}   An object of class \code{S4}. Access internal
objects from slots using @.

\code{BC07_preproc@alt.qc}   A 79 x 106 matrix of pre-processed alternative
read counts from the bulk and single-cell data for BC07. The row names denote
the chromosomal position of the point mutations. The column names denote the
bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood samples according to the
above labels from the SNV callset of patient BC07.

\code{BC07_preproc@annovar}   A 79 x 39 dataframe containing the
functional annotation of the somatic variants for BC07. For details regarding
the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}.

\code{BC07_preproc@featurecounts.qc}   A 63677 x 101 matrix of
pre-processed mapped reads from single-cell gene expression data. The row
names denote the gene names. The column names from the third column onward
denote the scRNA-seq samples from the gene expression data for patient BC07.

\code{BC07_preproc@mut}   A 79 x 5 dataframe containing the chromosomal
positions of point mutations for BC07. The columns consist of: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coincides with
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{BC07_preproc@ref.qc}   A 79 x 106 matrix of pre-processed
reference read counts from the bulk and single-cell data for BC07. The
row names denote the chromosomal position of the point mutations. The
column names denote the bulk RNA-seq, scRNA-seq, bulk DNA WES, and blood
samples from the total read counts of patient BC07.

\code{BC07_preproc@summary.qc}   A 14 x 101 matrix of the
\code{featureCounts} output summary that reports assignment of alignments
to genomic features. The row names denote the assignment of the alignments.
The column names denote the scRNA-seq samples from the gene expression data
for patient BC07.
}
\usage{
BC07_preproc
}
\description{
Pre-processed single-cell RNA sequencing (scRNA-seq), bulk RNA sequencing
(bulk RNA-seq), and bulk DNA whole exome sequencing data (bulk DNA WES)
from one patient (BC07) of 11 individuals (BC01-BC11) with primary breast
cancer \insertCite{Chung2017}{Canopy2}. The single-nucleotide variant (SNV)
pipeline utilized to pre-process the data is described in the main text and
supplement. In the main study, the authors sequenced 549 primary breast
cancer cells and matched bulk tumors and/or pooled cells from 11 patients,
including two lymph node metastases (BC03LN, BC07LN). Their final dataset
after filtering included 515 single cells and 14 bulk samples. \if{html}{\out{<br />}}
GEO Accession Number: GSE75688
}
\details{
\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (pre-processed)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   BC07 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab Metastatic lymph node \tab 1 \tab 1 (pooled bulk RNA) \tab 53 (52) \cr
    \tab Primary tumor \tab 1 \tab 1 (tumor bulk RNA) \tab 51 (49) \cr
    \tab (total) \tab 3 \tab 2 \tab 104 (101) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
