% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_post_v2.R
\name{canopy_post_v2}
\alias{canopy_post_v2}
\title{Internal function}
\usage{
canopy_post_v2(
  sampchain,
  projectname,
  K,
  numchain,
  burnin,
  thin,
  optK,
  C = NULL,
  post.config.cutoff = NULL
)
}
\description{
Modify Canopy function to remove cutoff value for posterior probabilities.
}
\keyword{internal}
