% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_fun.R
\name{dic_fun}
\alias{dic_fun}
\title{Spiegelhalter DIC}
\usage{
dic_fun(
  logPost.mean,
  Z.mean,
  Ps.mean,
  Pb.mean,
  Rb,
  Xb,
  Rs,
  Xs,
  alpha,
  beta,
  kappa,
  tau
)
}
\arguments{
\item{logPost.mean}{a numeric representing the mean (across all iterations)
of the log posteriors from the best chain}

\item{Z.mean}{a numeric representing the mean (across all iterations) of the
clonal configuration matrices, Z, from the best chain}

\item{Ps.mean}{a numeric representing the mean (across all iterations) of the
cell-to-clone assignment matrices, Ps, from the best chain}

\item{Pb.mean}{a numeric representing the mean (across all iterations) of the
sample-to-clone assignment matrices, Pb, from the best chain}

\item{Rb}{\eqn{M} (mutations) x \eqn{S} (bulk samples) matrix of bulk
alternative read counts. Rownames must match those of argument \code{Rs}.}

\item{Xb}{\eqn{M} (mutations) x \eqn{S} (bulk samples) matrix of bulk total
(benign + mutated) read counts. Rownames must match those of argument
\code{Xs}.}

\item{Rs}{\eqn{M} (mutations) x \eqn{N} (cells) matrix of single-cell
alternative read counts. Rownames must match those of argument \code{Rb}.}

\item{Xs}{\eqn{M} (mutations) x \eqn{N} (cells) matrix of single-cell
total (benign + mutated) read counts. Rownames must match those of argument
\code{Xb}.}

\item{alpha}{numeric vector of positive values with length \eqn{M}
(number of mutations) representing the gene activation rates.}

\item{beta}{numeric vector of positive values with length \eqn{M}
(number of mutations) representing the gene deactivation rates.}

\item{kappa}{a positive value used in the computation of the
sequencing error defined as \eqn{\kappa/(\kappa+\tau)}.}

\item{tau}{a positive value used in the computation of the sequencing error
defined as \eqn{\kappa/(\kappa+\tau)}.}
}
\value{
A numeric representing the deviance information criterion (DIC) according
to the Spiegelhalter definition \insertCite{Spiegelhalter2002}{canopy2}.
}
\description{
Produce the optimal configuration of the phylogenetic tree by using the
deviance information criterion (DIC) \insertCite{Spiegelhalter2002}{canopy2}.
}
\details{
According to the Spiegelhalter definition
\insertCite{Spiegelhalter2002}{canopy2}, the deviance information criterion
is calculated as

\deqn{DIC = p_D + \overline{D(\theta)}}

where \eqn{D(\theta)=-2log(p(y|\theta)) + C} is the deviance and
\eqn{p_D = \overline{D(\theta)} - D(\bar{\theta})} is the effective
number of parameters. Here \eqn{\theta} denotes the unknown parameters,
\eqn{y} denotes the observed data, and \eqn{\bar{\theta}} denotes the mean
of \eqn{\theta}. In calculations that compare models, \eqn{C} cancels out so
it can typically be ignored.
}
\examples{

# Simulate read counts
sims.out<-simulate_data(N=15, S=2, M=5, alpha=0.1, beta=1, kappa=1, tau=999,
                       Ktrue=4, b.mindepth=30, b.maxdepth=50, sc.mindepth=80,
                       sc.maxdepth=120, scale=300, seed=8675309)

# Run Canopy2 to get list of phylogenetic trees corresponding to all chains
# and all subclones
get.trees.out<-get_trees(Rs=sims.out$Rs, Rb=sims.out$Rb,
                        Xs=sims.out$Xs, Xb=sims.out$Xb,
                        alpha=sims.out$alpha, beta=sims.out$beta,
                        kappa=1, tau=999, Klist=3:5,
                        niter=5000, nchains=3, thin=10, pburn=0.1, seed=8675309)

# Get best tree across all chains and subclones via DIC
best.tree.out<-get_best_tree(get.trees.out)

# Compute absolute reconstruction error for Z
are_z_fun(true.Z=sims.out$true.tree$Z, inferred.Z=best.tree.out$tree$Z)

}
\references{
{
  \insertAllCited{}
}
}
