% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metric_fun.R
\name{dist_metric_fun}
\alias{dist_metric_fun}
\title{Distance Metric}
\usage{
dist_metric_fun(truth, inferred)
}
\arguments{
\item{truth}{a matrix of binary or continuous (positive) values that
represents the ground truth}

\item{inferred}{a matrix of binary or continuous (positive) values that
represents the inference}
}
\value{
A list containing: \code{dist.metric}, a numeric corresponding to the minimum
distance metric and \code{mat.perm.min}, a matrix representing the
permutation corresponding to the minimum distance metric.
}
\description{
Compute the absolute distance between a true matrix and a permuted matrix.
}
\details{
Please refer to Algorithm S3 in the manuscript text for details regarding
how the distance metric is computed. In brief, the distance metric is
computed by i) generating all columnwise permutations of whichever matrix
(inferred or true) has more columns and ii) for each permutation, computing
the distance metric as defined by Algorithm S3.
}
\examples{
# Simulate read counts
sims.out<-simulate_data(N=15, S=2, M=5, alpha=0.1, beta=1, kappa=1, tau=999,
                       Ktrue=4, b.mindepth=30, b.maxdepth=50, sc.mindepth=80,
                       sc.maxdepth=120, scale=300, seed=8675309)

# Run Canopy2 to get list of phylogenetic trees corresponding to all chains
# and all subclones
get.trees.out<-get_trees(Rs=sims.out$Rs, Rb=sims.out$Rb,
                        Xs=sims.out$Xs, Xb=sims.out$Xb,
                        alpha=sims.out$alpha, beta=sims.out$beta,
                        kappa=1, tau=999, Klist=3:5,
                        niter=5000, nchains=3, thin=10, pburn=0.1, seed=8675309)

# Get best tree across all chains and subclones via DIC
best.tree.out<-get_best_tree(get.trees.out)

# Distance between true and inferred clonal configuration matrices, Z
dist.Z<-dist_metric_fun(truth=sims.out$true.tree$Z, inferred=best.tree.out$tree$Z)

# Print the minimum distance metric
dist.Z$dist.metric

# Print the permutation corresponding to the minimum distance metric
dist.Z$mat.perm.min
}
