% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_data.R
\docType{data}
\name{GBM10_postproc}
\alias{GBM10_postproc}
\title{Post-processed breast cancer (BC) data from patient GBM10}
\format{
Labels:
\itemize{
\item Bulk samples (RNA-seq): \if{html}{\out{<br />}}
_EGAR00001456512_GBM10_1.RNA, \if{html}{\out{<br />}}
_EGAR00001456513_GBM10_2.RNA, \if{html}{\out{<br />}}
_EGAR00001456514_GBM10_3.RNA
\item Single-cell samples (RNA-seq): \if{html}{\out{<br />}}
_EGAR00001455959_GBM10_1_SC06.RNA to _EGAR00001456043_GBM10_2_SC94.RNA
\item Blood (normal) samples: \if{html}{\out{<br />}}
_EGAR00001456508_GBM10.DNA_B
\item Bulk samples (DNA WES): \if{html}{\out{<br />}}
_EGAR00001456471_GBM10_1.DNA, \if{html}{\out{<br />}}
_EGAR00001456473_GBM10_2.DNA, \if{html}{\out{<br />}}
_EGAR00001456472_GBM10_3.DNA
}

\code{GBM10_postproc}   An object of class \code{S4}. Access internal
objects from slots using @.

\code{GBM10_postproc@annovar.qc}   A 8 x 39 dataframe containing the
post-processed functional annotations of the somatic variants. For details
regarding the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}. Only those
annotations corresponding to mutations with at least one non-zero count
across all cells were retained.

\code{GBM10_postproc@featurecounts.qc}   A 63677 x 73 matrix of
post-processed mapped reads from the single-cell gene expression data.
The row names denote the gene names. The column names from the third column
onward denote the scRNA-seq samples from the gene expression data for patient
GBM10. Only those cells that have at least one non-zero count across genes
were retained.

\code{GBM10_postproc@param.est}   A list of 5 elements: \code{alpha},
\code{beta}, \code{scale}, \code{id.g}, and \code{pct.estimable}. The first
three are of length equal to the number of mutations where \code{alpha} denotes the
mutation-specific gene activation rates, \code{beta} denotes the mutation-specific
gene deactivation rates, and scale denotes the mutation-specific
transcription rates. The ids in \code{id.g} denote the positions corresponding
to the mutations that were estimable (returning non-NA or non-negative values),
and \code{pct.estimable} denotes the percentage that were estimable.

\code{GBM10_postproc@Rb}   A 8 x 3 matrix of post-processed alternative
read counts from the bulk DNA WES data. The row names denote the chromosomal
position of the point mutations. The column names denote the bulk samples
according to the above labels from the SNV callset of patient GBM10.
Post-processing involved removing mutations (rows) with zero alternative
read counts across all cells from the single-cell data.

\code{GBM10_postproc@Rs}   A 8 x 71 matrix of post-processed alternative
read counts from the scRNA-seq data. The row names denote the chromosomal
position of the point mutations. The column names denote the scRNA-seq samples
according to the above labels from the SNV callset of patient GBM10.
Post-processing involved i) removing mutations (rows) with zero alternative
read counts across all cells and ii) removing cells (columns) with zero total
(benign + mutated) read counts across all mutations.

\code{GBM10_postproc@Xb}   A 8 x 3 matrix of post-processed total read
counts (benign + mutated) from the bulk DNA WES data. The row names denote
the chromosomal position of the point mutations. The column names denote the
bulk samples according to the above labels from the SNV callset of patient
GBM10. Post-processing involved removing mutations (rows) with zero
alternative read counts across all cells from the single-cell data.

\code{GBM10_postproc@Xs}   A 8 x 71 matrix of post-processed total read
counts (benign + mutated) from the scRNA-seq data. The row names denote the
chromosomal position of the point mutations. The column names denote the
scRNA-seq samples according to the above labels from the SNV callset of
patient GBM10. Post-processing involved i) removing mutations (rows) with zero
alternative read counts across all cells and ii) removing cells (columns)
with zero total (benign + mutated) read counts across all mutations.
}
\usage{
GBM10_postproc
}
\description{
Some further data munging of GBM10_preproc to get GBM10_postproc, which contains
the finalized read counts for Canopy2 functions and vignette.\if{html}{\out{<br />}}
Accession number: EGAS00001001880
}
\details{
\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (post-processed)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   GBM10 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab GBM10_1 \tab 1 \tab 1 \tab 31 (28) \cr
    \tab GBM10_2 \tab 1 \tab 1 \tab 54 (43) \cr
    \tab GBM10_3 \tab 1 \tab 1 \tab 0 \cr
    \tab (total) \tab 4 \tab 3 \tab 85 (71) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
