% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_data.R
\docType{data}
\name{GBM2}
\alias{GBM2}
\title{Glioblastoma (GBM) Data from Patient GBM2}
\format{
Labels:
\itemize{
\item Single-cell samples (RNA-seq): _EGAR00001456045_GBM2_1_SC02.RNA to
_EGAR00001456180_GBM2_2_SC95.RNA
\item Blood (normal) samples: _EGAR00001456500_GBM2.DNA_B
\item Bulk samples (DNA WES): _EGAR00001456446_GBM2_1.DNA, _EGAR00001456447_GBM2_2.DNA
}

\code{GBM2}   An object of class \code{S4}. Access internal objects
from slots using @.

\code{GBM2@alt.qc}   A matrix with 124 columns and 653 rows. The column
names denote the scRNA-seq samples (\code{.RNA}), bulk DNA WES samples
(\code{.DNA}), and blood samples (\code{.DNA_B}) from the SNV callset of
patient GBM2. The row names denote the chromosomal position of the point
mutations.

\code{GBM2@annovar}   A dataframe with 39 columns and 653 rows.
For details regarding the column headers, we refer readers to the extensive
\href{https://annovar.openbioinformatics.org/}{ANNOVAR} documentation and the
original research article \insertCite{Wang2010}{Canopy2}.

\code{GBM2@featurecounts.qc}   A matrix with 121 columns and 63677 rows.
The column names denote the scRNA-seq samples from the gene expression data
for patient GBM2. The row names denote the gene names.

\code{GBM2@mut}   A data frame with five variables: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coincideswith
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{GBM2@ref.qc}   A matrix with 124 columns and 653 rows. The
column names denote the scRNA-seq samples (\code{.RNA}), bulk DNA WES samples
(\code{.DNA}), and blood samples (\code{.DNA_B}) from the total read counts
of patient GBM2. The row names denote the chromosomal position of the point
mutations.

\code{GBM2@summary.qc}   A matrix with 121 columns and 14 rows. The
column names denote the scRNA-seq samples from the gene expression data for
patient GBM2. The row names denote the assignment of the alignments.
}
\usage{
GBM2
}
\description{
QC'd single-cell RNA sequencing (scRNA-seq) and bulk DNA whole exome
sequencing data (bulk DNA WES) from one patient (GBM2) selected from
52 individuals with glioblastoma (GBM). The authors characterized genomic and
expression profiles across 127 multisector or longitudinal specimens with
matched tumor/normal samples \insertCite{Lee2017}{Canopy2}. The
single-nucleotide variant (SNV) pipeline utilized to QC the data is described
in Figure S20 of \insertCite{Wang2021}{Canopy2}. Tumor samples
were acquired based on 4 categories: 1) locally adjacent tumors,
2) multifocal/multicentric tumors, 3) 5-ALA (+/-) tumors and
4) Longitudinal tumors. Accession number: EGAS00001001880
}
\details{
\code{GBM2@alt.qc}   Preprocessed alternative read counts from bulk DNA
WES and scRNA-seq for GMB2.

\code{GBM2@annovar}   Functional annotation of somatic variants for GBM2.

\code{GBM2@featurecounts.qc}   Mapped reads from single-cell gene
expression data for GBM2.

\code{GBM2@mut}   Chromosomal positions of point mutations for GBM2.

\code{GBM2@ref.qc}   Preprocessed reference read counts from bulk DNA WES and
single-cell RNA-seq for GBM2.

\code{GBM2@summary.qc}   \code{featureCounts} output summary that reports
assignment of alignments to genomic features

\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (QC)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   GBM2 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab GBM2_1 \tab 1 \tab 0 \tab 74 (67) \cr
    \tab GBM2_2 \tab 1 \tab 0 \tab 63 (54) \cr
    \tab (total) \tab 3 \tab 0 \tab 137 (121) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
