% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_data.R
\docType{data}
\name{GBM9}
\alias{GBM9}
\title{Glioblastoma (GBM) Data from Patient GBM9}
\format{
Labels:
\itemize{
\item Bulk samples (RNA-seq): _EGAR00001456543_GBM9_1.RNA (initial tumor),
_EGAR00001456544_GBM9_2.RNA (initial tumor),
_EGAR00001456545_GBM9_R1.RNA (recurrent tumor),
_EGAR00001456546_GBM9_R2.RNA (recurrent tumor)
\item Single-cell samples (RNA-seq): _EGAR00001456181_GBM9_1_SC07.RNA to
_EGAR00001456269_GBM9_2_SC94.RNA (initial tumor)
&
_EGAR00001456270_GBM9_R1_SC05.RNA to
_EGAR00001456313_GBM9_R1_SC93.RNA (recurrent tumor)
\item Blood (normal) samples: _EGAR00001456508_GBM9.DNA_B
\item Bulk samples (DNA WES): _EGAR00001456467_GBM9_1.DNA (initial tumor),
_EGAR00001456468_GBM9_2.DNA (initial tumor),
_EGAR00001456469_GBM9_R1.DNA (recurrent tumor),
_EGAR00001456470_GBM9_R2.DNA (recurrent tumor)
}

\code{GBM9}   An object of class \code{S4}. Access internal objects
from slots using @.

\code{GBM9@alt.qc}   A matrix with 135 columns and 143 rows. The column
names denote the scRNA-seq and bulk RNA-seq samples (\code{.RNA}),
bulk DNA WES samples (\code{.DNA}), and blood samples (\code{.DNA_B}) from
the SNV callset of patient GBM9. The row names denote the chromosomal
position of the point mutations.

\code{GBM9@annovar}   A dataframe with 39 columns and 143 rows.
For details regarding the column headers, we refer readers to the extensive
\href{https://annovar.openbioinformatics.org/}{ANNOVAR} documentation and the
original research article \insertCite{Wang2010}{Canopy2}.

\code{GBM9@featurecounts.qc}   A matrix with 126 columns and 63677 rows.
The column names denote the scRNA-seq samples from the gene expression data
for patient GBM9. The row names denote the gene names.

\code{GBM9@mut}   A data frame with five variables: the chromosome
number \code{Chr}, the starting position of the point mutation \code{Start},
the ending position of the point mutation \code{End} (coincideswith
\code{Start} because the mutation is a single nucleotide variant), the
reference allele \code{Ref}, and the alternate allele \code{Alt}.

\code{GBM9@ref.qc}   A matrix with 135 columns and 143 rows. The
column names denote the scRNA-seq and bulk RNA-seq samples (\code{.RNA}),
bulk DNA WES samples (\code{.DNA}), and blood samples (\code{.DNA_B})
from the total read counts of patient GBM9. The row names denote the
chromosomal position of the point mutations.

\code{GBM9@summary.qc}   A matrix with 126 columns and 14 rows. The
column names denote the scRNA-seq samples from the gene expression data for
patient GBM9. The row names denote the assignment of the alignments.
}
\usage{
GBM9
}
\description{
QC'd single-cell RNA sequencing (scRNA-seq), bulk RNA sequencing (bulk RNA-seq)
and bulk DNA whole exome sequencing data (bulk DNA WES) from one patient (GBM9)
selected from 52 individuals with glioblastoma (GBM). The authors characterized
genomic and expression profiles across 127 multisector or longitudinal specimens
with matched tumor/normal samples \insertCite{Lee2017}{Canopy2}. The
single-nucleotide variant (SNV) pipeline utilized to QC the data is described
in Figure S20 of \insertCite{Wang2021}{Canopy2}. Tumor samples
were acquired based on 4 categories: 1) locally adjacent tumors,
2) multifocal/multicentric tumors, 3) 5-ALA (+/-) tumors and
4) Longitudinal tumors. Accession number: EGAS00001001880
}
\details{
\code{GBM9@alt.qc}   Preprocessed alternative read counts from bulk DNA
WES and scRNA-seq for GMB9.

\code{GBM9@annovar}   Functional annotation of somatic variants for GBM9.

\code{GBM9@featurecounts.qc}   Mapped reads from single-cell gene
expression data for GBM9.

\code{GBM9@mut}   Chromosomal positions of point mutations for GBM9.

\code{GBM9@ref.qc}   Preprocessed reference read counts from bulk DNA WES and
single-cell RNA-seq for GBM9.

\code{GBM9@summary.qc}   \code{featureCounts} output summary that reports
assignment of alignments to genomic features

\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (QC)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   GBM9 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab GBM9_1 (initial tumor) \tab 1 \tab 1 \tab 29 (25) \cr
    \tab GBM9_2 (initial tumor) \tab 1 \tab 1 \tab 60 (58) \cr
    \tab GBM9_R1 (recurrent tumor) \tab 1 \tab 1 \tab 44 (43) \cr
    \tab GMB9_R2 (recurrent tumor) \tab 1 \tab 1 \tab 0 \cr
    \tab (total) \tab 5 \tab 4 \tab 133 (126) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
