% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_data.R
\docType{data}
\name{GBM9_postproc}
\alias{GBM9_postproc}
\title{Post-processed breast cancer (BC) data from patient GBM9}
\format{
Labels:
\itemize{
\item Bulk samples (RNA-seq): \if{html}{\out{<br />}}
_EGAR00001456543_GBM9_1.RNA (initial tumor), \if{html}{\out{<br />}}
_EGAR00001456544_GBM9_2.RNA (initial tumor), \if{html}{\out{<br />}}
_EGAR00001456545_GBM9_R1.RNA (recurrent tumor), \if{html}{\out{<br />}}
_EGAR00001456546_GBM9_R2.RNA (recurrent tumor)
\item Single-cell samples (RNA-seq): \if{html}{\out{<br />}}
_EGAR00001456181_GBM9_1_SC07.RNA to _EGAR00001456269_GBM9_2_SC94.RNA (initial tumor) \if{html}{\out{<br />}}
& \if{html}{\out{<br />}}
_EGAR00001456270_GBM9_R1_SC05.RNA to \if{html}{\out{<br />}}
_EGAR00001456313_GBM9_R1_SC93.RNA (recurrent tumor)
\item Blood (normal) samples: _EGAR00001456508_GBM9.DNA_B
\item Bulk samples (DNA WES): \if{html}{\out{<br />}}
_EGAR00001456467_GBM9_1.DNA (initial tumor), \if{html}{\out{<br />}}
_EGAR00001456468_GBM9_2.DNA (initial tumor), \if{html}{\out{<br />}}
_EGAR00001456469_GBM9_R1.DNA (recurrent tumor), \if{html}{\out{<br />}}
_EGAR00001456470_GBM9_R2.DNA (recurrent tumor)
}

\code{GBM9_postproc}   An object of class \code{S4}. Access internal
objects from slots using @.

\code{GBM9_postproc@annovar.qc}   A 23 x 39 dataframe containing the
post-processed functional annotations of the somatic variants. For details
regarding the column headers, we refer readers to the extensive
ANNOVAR \href{https://annovar.openbioinformatics.org/}{documentation} and the
original research article \insertCite{Wang2010}{Canopy2}. Only those
annotations corresponding to mutations with at least one non-zero count
across all cells were retained.

\code{GBM9_postproc@featurecounts.qc}   A 63677 x 127 matrix of
post-processed mapped reads from the single-cell gene expression data.
The row names denote the gene names. The column names from the third column
onward denote the scRNA-seq samples from the gene expression data for patient
GBM9. Only those cells that have at least one non-zero count across genes
were retained.

\code{GBM9_postproc@param.est}   A list of 5 elements: \code{alpha},
\code{beta}, \code{scale}, \code{id.g}, and \code{pct.estimable}. The first
three are of length equal to the number of mutations where \code{alpha} denotes the
mutation-specific gene activation rates, \code{beta} denotes the mutation-specific
gene deactivation rates, and \code{scale} denotes the mutation-specific
transcription rates. The ids in \code{id.g} denote the positions corresponding
to the mutations that were estimable (returning non-NA or non-negative values),
and \code{pct.estimable} denotes the percentage that were estimable.

\code{GBM9_postproc@Rb}   A 18 x 4 matrix of post-processed alternative
read counts from the bulk DNA WES data. The row names denote the chromosomal
position of the point mutations. The column names denote the bulk samples
according to the above labels from the SNV callset of patient GBM9.
Post-processing involved removing mutations (rows) with zero alternative
read counts across all cells from the single-cell data.

\code{GBM9_postproc@Rs}   A 18 x 125 matrix of post-processed alternative
read counts from the scRNA-seq data. The row names denote the chromosomal
position of the point mutations. The column names denote the scRNA-seq samples
according to the above labels from the SNV callset of patient GBM9.
Post-processing involved i) removing mutations (rows) with zero alternative
read counts across all cells and ii) removing cells (columns) with zero total
(benign + mutated) read counts across all mutations.

\code{GBM9_postproc@Xb}   A 18 x 4 matrix of post-processed total read
counts (benign + mutated) from the bulk DNA WES data. The row names denote
the chromosomal position of the point mutations. The column names denote the
bulk samples according to the above labels from the SNV callset of patient
GBM9. Post-processing involved removing mutations (rows) with zero
alternative read counts across all cells from the single-cell data.

\code{GBM9_postproc@Xs}   A 18 x 125 matrix of post-processed total read
counts (benign + mutated) from the scRNA-seq data. The row names denote the
chromosomal position of the point mutations. The column names denote the
scRNA-seq samples according to the above labels from the SNV callset of
patient GBM9. Post-processing involved i) removing mutations (rows) with zero
alternative read counts across all cells and ii) removing cells (columns)
with zero total (benign + mutated) read counts across all mutations.
}
\usage{
GBM9_postproc
}
\description{
Some further data munging of GBM9_preproc to get GBM9_postproc, which contains
the finalized read counts for Canopy2 functions and vignette. \if{html}{\out{<br />}}
Accession number: EGAS00001001880
}
\details{
\tabular{lllll}{
   \emph{\strong{Patient}} \tab \emph{\strong{Subtype}} \tab \emph{\strong{# bulk DNA WES}} \tab \emph{\strong{# bulk RNA-seq}} \tab \emph{\strong{# scRNA-seq (post-processed)}} \cr
   ----- \tab ----- \tab ----- \tab ----- \tab ----- \cr
   GBM9 \tab Blood \tab 1 \tab 0 \tab 0 \cr
    \tab GBM9_1 (initial tumor) \tab 1 \tab 1 \tab 29 (25) \cr
    \tab GBM9_2 (initial tumor) \tab 1 \tab 1 \tab 60 (58) \cr
    \tab GBM9_R1 (recurrent tumor) \tab 1 \tab 1 \tab 44 (43) \cr
    \tab GMB9_R2 (recurrent tumor) \tab 1 \tab 1 \tab 0 \cr
    \tab (total) \tab 5 \tab 4 \tab 133 (125) \cr
}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
