% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_tree_sims.R
\name{get_best_tree_sims}
\alias{get_best_tree_sims}
\title{Best Phylogeny}
\usage{
get_best_tree_sims(
  truetree,
  get.trees.out,
  Ktrue,
  criterion = c("AIC", "BIC", "DIC"),
  save.muts = F,
  save.plot = F,
  project = NULL,
  outpath = NULL
)
}
\arguments{
\item{truetree}{phylogenetic tree of class \code{phylo}.}

\item{get.trees.out}{sample of phylogenetic trees output from function
\code{get_trees()} of class \code{get_trees}.}

\item{Ktrue}{true number of subclones (1 or more).}

\item{criterion}{information criterion used for model selection and must be
one of \code{"AIC"}, \code{"BIC"}, and \code{"DIC"}.}

\item{save.muts}{a logical indicating whether a text file containing all
mutations along the branches should be saved. If \code{TRUE}, must specify
\code{project} and \code{outpath}.}

\item{save.plot}{a logical indicating whether plots of both the true and
best tree should be saved. If \code{TRUE}, must specify \code{project} and
\code{outpath}.}

\item{project}{a project name to add to the filename for output generated
from \code{save.muts} and \code{save.plot}.}

\item{outpath}{a location at which to save output generated from
\code{save.muts} and \code{save.plot}.}
}
\description{
Produce the optimal phylogenetic tree from simulations based on a
user-supplied information criterion.
}
