% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BM_fun.R
\name{get_BM_fun}
\alias{get_BM_fun}
\title{Map chromosome positions to HGNC symbols}
\usage{
get_BM_fun(reads)
}
\arguments{
\item{reads}{a matrix with row names of the form chr(num):(pos) where num is a
positive integer representing the chromosome number and pos is the positive
integer representing the variant position, e.g., chr1:75190429, and
column names representing the samples (i.e., single-cell or bulk).}
}
\value{
The original matrix with an extra column (appended as the first column) that
contains the HGNC symbols corresponding to the chromosome positions.
}
\description{
Use biomaRt to map chromosome positions to HGNC symbols and return original
read counts with appended HGNC symbols.
}
\examples{
#Load post-processed data for patient GBM10
data("GBM10_postproc")

# Map chromosome positions to gene IDs
Rs<-get_BM_fun(GBM10_postproc@Rs)

head(Rs)

}
