% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_burstiness_bpsc.R
\name{get_burstiness_bpsc}
\alias{get_burstiness_bpsc}
\title{Estimate Bursting Kinetics with BPSC}
\usage{
get_burstiness_bpsc(featurecounts, N)
}
\arguments{
\item{featurecounts}{a matrix of scRNA-seq counts}

\item{N}{number of single cells}
}
\description{
Using the BPSC methodology \insertCite{Vu2016}{Canopy2},
estimate two parameters for bursting kinetics: the gene activation rate
(alpha) and the gene deactivation rate (beta) from the gene expression data.
}
\references{
{
  \insertAllCited{}
}
}
