% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_burstiness_scale.R
\name{get_burstiness_scale}
\alias{get_burstiness_scale}
\title{Estimate Bursting Kinetics with SCALE}
\usage{
get_burstiness_scale(featurecounts, N)
}
\arguments{
\item{featurecounts}{a matrix of scRNA-seq counts}

\item{N}{number of single cells}
}
\description{
Using the SCALE methodology \insertCite{Jiang2017}{Canopy2}, estimate two
parameters for bursting kinetics, the gene activation rate (alpha)
and the gene deactivation rate (beta), from the gene expression data.
}
\references{
{
  \insertAllCited{}
}
}
