% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnostics.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Diagnostic Plots}
\usage{
get_diagnostics(get.trees.out, project = NULL, outpath = NULL)
}
\arguments{
\item{get.trees.out}{sample of phylogenetic trees output from function
\code{get_trees()} of class \code{get_trees}.}

\item{project}{a project name to add to the filename for output generated
from \code{save.muts} and \code{save.plot}.}

\item{outpath}{a location at which to save output generated from
\code{save.muts} and \code{save.plot}.}
}
\description{
Plot posterior densities, trace, and lag autocorrelation factor (lag ACF).
Requires output from function \code{get_trees()}.
}
