% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expression.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Secondary processing of gene expression data}
\usage{
get_gene_expression(featurecounts, build = 37)
}
\arguments{
\item{featurecounts}{a gene x sample matrix of pre-processed read counts from
gene expression data.}

\item{a}{positive integer corresponding to the Genome Reference Consortium
Human (GRCh) build. This is passed to the biomaRt::useEnsembl() function which can
currently only take build 37 (as of June 11, 2023).}
}
\value{
The original matrix with two extra columns (appended as the first two columns)
that contain the Ensembl gene IDs and corresponding HGNC symbols.
}
\description{
Secondary processing of gene expression data
}
\examples{
#Load pre-processed data for patient GBM10
data("GBM10_preproc")
featurecounts.qc0<-GBM10_preproc@featurecounts.qc

# Get HGNC symbol associated with ensembl gene id
featurecounts.qc<-get_gene_expression(featurecounts=featurecounts.qc0,
                                      build=37)

head(featurecounts.qc)

}
