% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Internal function}
\usage{
get_os()
}
\description{
Detect the user's operating system (OS).
}
\details{
See this link for further details:
https://www.r-bloggers.com/identifying-the-os-from-r/
}
\examples{
\dontrun{
get_os()
}
}
\keyword{internal}
