% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{get_trees}
\alias{get_trees}
\title{Sample space of phylogenetic trees}
\usage{
get_trees(
  Rs,
  Rb,
  Xs,
  Xb,
  alpha,
  beta,
  kappa,
  tau,
  Klist,
  niter,
  nchains,
  thin,
  pburn,
  seed
)
}
\arguments{
\item{Rs}{matrix with rownames matching those of argument \code{Rb}}

\item{Rb}{matrix with rownames matching those of argument \code{Rs}}

\item{Xs}{matrix with rownames matching those of argument \code{Xb}}

\item{Xb}{matrix with rownames matching those of argument \code{Xs}}

\item{alpha}{numeric vector of values greater than 0 with length \eqn{K}
(number of subclones) representing the gene activation rate}

\item{beta}{numeric vector of values greater than 0 with length \eqn{K}
(number of subclones) representing the gene deactivation rate}

\item{kappa}{a positive value used in the computation of the
sequencing error defined as \eqn{\kappa/(\kappa+\tau)}}

\item{tau}{a positive value used in the computation of the sequencing error
defined as \eqn{\kappa/(\kappa+\tau)}}

\item{Klist}{numeric vector containing the possible numbers of subclones}

\item{niter}{number of iterations of MCMC}

\item{nchains}{number of chains for MCMC}

\item{thin}{a number representing the increment at which to store MCMC
output. For example, if \code{thin=10} then every 10th iteration is stored.}

\item{pburn}{a decimal denoting the percentage of burn-in to store}

\item{seed}{a state (positive integer) to set the random number generation}
}
\description{
Produce a list of phylogenetic trees in a sample space of size
number of chains x number of subclones. Sampling is performed using
Metropolis-within-Gibbs. Output is coded as class \code{get_trees}.
}
