% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZ.R
\name{getZ}
\alias{getZ}
\title{Generate the clonal configuration matrix, Z}
\usage{
getZ(tree)
}
\arguments{
\item{tree}{an object of class \code{pyhlo} representing the phylogenetic
tree.}
}
\value{
A M (mutations) x K (clones) binary clonal configuration matrix, Z, that
indicates which mutations belong to which clones.
}
\description{
Derive the binary clonal configuration matrix, Z, from a phylogenetic
tree.
}
\examples{
# Generate a tree
K<-3
text = paste0(paste0(paste(paste0("(", 1:(K - 1), ","),
                           collapse = ""), K), paste(rep(")", (K - 1)), collapse = ""), ";")
tree <- ape::read.tree(text = text); rm(text)

# Generate point mutations (single-nucleotide varaints) along the tree branches
tree$snv<-initialsnv(tree, snv.name=paste("mut",1:5))

# Get the Z matrix from tree and SNVs
tree$Z<-getZ(tree)

tree$Z

}
