% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tree.R
\name{initialize_tree}
\alias{initialize_tree}
\title{Internal function}
\usage{
initialize_tree(seedling, K, Rs, S, Rb, N, Xb, Xs, alpha, beta, kappa, tau)
}
\description{
Initialize phylogenetic tree.
}
\keyword{internal}
