% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialsnv.R
\name{initialsnv}
\alias{initialsnv}
\title{Generate SNVs}
\usage{
initialsnv(tree, snv.name)
}
\arguments{
\item{tree}{an object of class \code{pyhlo} representing the phylogenetic
tree.}

\item{snv.name}{an object of class \code{character} containing the names
of the mutations.}
}
\value{
A matrix with five columns: \code{snv} (the SNV number), \code{snv.strt.node}
(the start node number), \code{snv.end.node} (the end node number), and
\code{snv.edge} (the edge number). Rownames correspond to mutations.
}
\description{
Generate point mutations (single-nucleotide variants, SNVs) along the tree
branches.
}
\examples{
# Generate a tree
K<-3
text = paste0(paste0(paste(paste0("(", 1:(K - 1), ","),
                           collapse = ""), K), paste(rep(")", (K - 1)), collapse = ""), ";")
tree <- ape::read.tree(text = text); rm(text)

# Generate point mutations (single-nucleotide varaints) along the tree branches
tree$snv<-initialsnv(tree, snv.name=paste("mut",1:5))

}
