% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Z.R
\name{plot_Z}
\alias{plot_Z}
\title{Plot Z}
\usage{
plot_Z(tree)
}
\arguments{
\item{tree}{an object of class \code{pyhlo} representing the phylogenetic
tree.}
}
\value{
A plot of class \code{ggtree} that is automatically plotted to the graphics
device.
}
\description{
Plot the clonal tree configuration matrix, Z.
}
\details{
The plotted tree is arranged upright such that time runs vertically from
the root to the tips of the tree. The leftmost branch corresponds to the
population of normal cells, while the remaining branches correspond to the
mutant cells. Each red text box, other than the root, denotes a cluster of
point mutations (single-nucleotide variants). For example, "M1: 2 SNVs" would
imply that cluster M1 contains two point mutations, perhaps mutations in genes
ABC and XYZ. Finally, the tips of each tree are labeled with the percentage
of cells assigned to each subclone.
}
\examples{
# Simulate read counts
sims.out<-simulate_data(N=30, S=2, M=5, alpha=0.1, beta=1.0, kappa=1, tau=999,
                        Ktrue=4, b.mindepth=30, b.maxdepth=50, sc.mindepth=80,
                        sc.maxdepth=120, scale=300, seed=8675309)

N <- ncol(sims.out$Rs) # Number of cells
S <- ncol(sims.out$Rb) # Number of bulk samples
M <- nrow(sims.out$Rs) # Number of point mutations

# Generate a tree
K<-4
text = paste0(paste0(paste(paste0("(", 1:(K - 1), ","),
             collapse = ""), K), paste(rep(")", (K - 1)), collapse = ""), ";")
tree <- ape::read.tree(text = text); rm(text)

# Generate point mutations (single-nucleotide varaints) along the tree branches
tree$snv<-initialsnv(tree, rownames(sims.out$Rs))

# Get the Z matrix from tree and snv
tree$Z <- getZ(tree)

# Generate the proportion matrices
# Bulk samples
Pb<-t(DirichletReg::rdirichlet(S, alpha=rep(1/K,K)))
rownames(Pb)<-paste0('clone',1:K)
colnames(Pb)<-colnames(sims.out$Rb)

# Single cells
Ps<-stats::rmultinom(N, 1, prob = rep(1/K, K))
rownames(Ps)<-paste0('clone',1:K)
colnames(Ps)<-colnames(sims.out$Rs)
tree$Pb<-Pb; rm(Pb)
tree$Ps<-Ps; rm(Ps)

# Plot Z
plot_Z(tree)

}
