% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBetaPois.R
\name{rBetaPois}
\alias{rBetaPois}
\title{Generate random variates from a beta-Poisson distribution}
\usage{
rBetaPois(n, alpha, beta, scale = 1)
}
\arguments{
\item{n}{The number of random values to return}

\item{alpha}{A positive value, the alpha shape parameter of the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing.}

\item{beta}{A positive value, the beta shape parameter of the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing.}

\item{scale}{A positive value, the scaling parameter for the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing.}
}
\value{
A vector of the random variates
}
\description{
Generate random variates from a beta-Poisson distribution
}
\examples{
set.seed(8675309)
bp.vec=rBetaPois(100,0.1,1.0,3000)
hist(bp.vec, prob=T)

}
