% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampPb.R
\name{sampPb}
\alias{sampPb}
\title{Internal function}
\usage{
sampPb(tree, ...)
}
\description{
Sample the bulk sample to clone assignment matrix, Pb, using Metropolis
Hastings (an accept reject algorithm). See Alg. 1 in the main text for
further details.
}
\examples{
\dontrun{
sim.tree<-initialize_tree(seedling=8675309)
sampPb(tree=sim.tree)
}
}
\keyword{internal}
