% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampPs.R
\name{sampPs}
\alias{sampPs}
\title{Internal function}
\usage{
sampPs(tree, ...)
}
\description{
Sample the single-cell to clone assignment matrix, Ps, using Metropolis
Hastings (an accept reject algorithm). See Alg. 1 in the main text for
further details.
}
\examples{
\dontrun{
sim.tree<-initialize_tree(seedling=8675309)
sampPs(tree=sim.tree)
}
}
\keyword{internal}
