% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampZ.R
\name{sampZ}
\alias{sampZ}
\title{Internal function}
\usage{
sampZ(tree, Rb, Xb, Rs, Xs, alpha, beta, kappa, tau, ...)
}
\description{
Sample the clonal configuration matrix, Z, using Metropolis Hastings (an
accept reject algorithm). See Alg. 1 in the main text for further details.
}
\examples{
\dontrun{
sim.tree<-initialize_tree(seedling=8675309)
sampZ(tree=sim.tree)
}
}
\keyword{internal}
