% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data for Canopy2}
\usage{
simulate_data(
  N,
  S,
  M,
  alpha,
  beta,
  kappa = 1,
  tau = 999,
  Ktrue,
  b.mindepth,
  b.maxdepth,
  sc.mindepth,
  sc.maxdepth,
  lam1 = 3000,
  lam2 = 1,
  seed = 1
)
}
\arguments{
\item{N}{number of single cells}

\item{S}{number of bulk samples}

\item{M}{number of mutations (single nucleotide variants)}

\item{alpha}{a positive integer representing the gene activation rate. This
number will be replicated to form a vector of length \eqn{K}.}

\item{beta}{a positive integer representing the gene deactivation rate. This
number will be replicated to form a vector of length \eqn{K}.}

\item{kappa}{a positive value used in the computation of the sequencing error
defined as \eqn{\kappa/(\kappa+\tau)}}

\item{tau}{a positive value used in the computation of the sequencing error
defined as \eqn{\kappa/(\kappa+\tau)}}

\item{Ktrue}{true number of subclones}

\item{b.mindepth}{a positive value that represents the minimum sequencing
depth of the bulk data}

\item{b.maxdepth}{a positive value that represents the maximum sequencing
depth of the bulk data}

\item{sc.mindepth}{a positive value that represents the minimum sequencing
depth of the single-cell data}

\item{sc.maxdepth}{a positive value that represents the maximum sequencing
depth of the single-cell data}

\item{lam1}{a positive value that represents the scaling factor for beta
distribution (the rate at which DNA transcribed into RNA)}

\item{lam2}{a positive value that represents the scaling factor for Poisson
distribution so that it can allow for fractional expression values}

\item{seed}{a state (positive integer) to set the random number generation}
}
\description{
Simulates data for Canopy2 from a beta-binomial distribution.
}
