% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clonal_composition.R
\name{get_clonal_composition}
\alias{get_clonal_composition}
\title{Get clonal composition}
\usage{
get_clonal_composition(tree)
}
\arguments{
\item{tree}{phylogenetic tree of class \code{phylo}.}
}
\value{
An M (mutations) x 1 matrix of the mutations assigned to each subclone.
The first subclone, "Clone 1" should not have any assigned mutations since it
corresponds to the normal subclone.
}
\description{
Returns the single nucleotide variants (SNVs) belonging to each subclone.
This function adapts the \code{getclonalcomposition} function from
\insertCite{Jiang2017}{canopy2} by removing any portions of the code related
to copy number alterations (CNAs).
}
\examples{
# Load post-processed data for patient GBM10
data("GBM10_postproc")

# Run Canopy2 to get list of phylogenetic trees corresponding to all chains
# and all subclones.
# Note: this is written to quickly compile for demonstration purposes, as we
# know the optimal number of subclones is 6. In practice, we would attempt a
# larger range of subclones (e.g., Klist=3:10) and a larger number of
# iterations and chains (e.g., niter=50000 and nchains=10).
get.trees.out<-get_trees(Rs=GBM10_postproc@Rs, Rb=GBM10_postproc@Rb,
                        Xs=GBM10_postproc@Xs, Xb=GBM10_postproc@Xb,
                        alpha=GBM10_postproc@param.est$alpha,
                        beta=GBM10_postproc@param.est$beta, kappa=1,
                        tau=999, Klist=5:7, niter=5000, nchains=5, thin=20,
                        pburn=0.2, seed=8675309)

# Examine diagnostic plots
get_diagnostics(get.trees.out)

# Get best tree across all chains and subclones via DIC
best.tree.out<-get_best_tree(get.trees.out)

# Return the single nucleotide variants (SNVs) belonging to each subclone
clonal.muts<-get_clonal_composition(tree=best.tree.out$tree)

clonal.muts

}
