% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{get_trees}
\alias{get_trees}
\title{Sample space of phylogenetic trees}
\usage{
get_trees(
  Rs,
  Rb,
  Xs,
  Xb,
  alpha,
  beta,
  kappa,
  tau,
  Klist,
  niter = 10000,
  nchains = 20,
  thin = 20,
  pburn = 0.5,
  ncores = NULL,
  seed = 8675309
)
}
\arguments{
\item{Rs}{\eqn{M} (mutations) x \eqn{N} (cells) matrix of single-cell
alternative read counts. Rownames must match those of argument \code{Rb}}

\item{Rb}{\eqn{M} (mutations) x \eqn{S} (bulk samples) matrix of bulk
alternative read counts. Rownames must match those of argument \code{Rs}}

\item{Xs}{\eqn{M} (mutations) x \eqn{N} (cells) matrix of single-cell
total (benign + mutated) read counts. Rownames must match those of argument
\code{Xb}.}

\item{Xb}{\eqn{M} (mutations) x \eqn{S} (bulk samples) matrix of bulk total
(benign + mutated) read counts. Rownames must match those of argument
\code{Xs}.}

\item{alpha}{numeric vector of positive values with length \eqn{M}
(number of mutations) representing the gene activation rates}

\item{beta}{numeric vector of positive values with length \eqn{M}
(number of mutations) representing the gene deactivation rates}

\item{kappa}{a positive value used in the computation of the
sequencing error defined as \eqn{\kappa/(\kappa+\tau)}}

\item{tau}{a positive value used in the computation of the sequencing error
defined as \eqn{\kappa/(\kappa+\tau)}}

\item{Klist}{numeric vector containing the possible numbers of subclones.}

\item{niter}{number of iterations of MCMC. Defaults to 10000.}

\item{nchains}{number of chains for MCMC. Defaults to 20.}

\item{thin}{a number representing the increment at which to store MCMC output.
For example, if \code{thin=10} then every 10th iteration is stored. Defaults
to 20.}

\item{pburn}{a decimal denoting the percentage of burn-in to store. Defaults
to 0.50 (50\\\%).}

\item{ncores}{the number of cores to use for parallel execution. If not
specified, defaults to one-half the number of cores detected by the
parallelly package.}

\item{seed}{a state (positive integer) to set the random number generation.
Defaults to 8675309.}
}
\value{
An object of class "get_trees" containing: \code{N}, the number of
single-cells, \code{S}, the number of bulk samples, \code{M}, the number of
mutations, \code{Rs}, the single-cell alternative read count matrix,
\code{Rb} the bulk alternative read count matrix, \code{Xs}, the single-cell
total (benign + mutated) read count matrix, \code{Xb}, the bulk total
(benign + mutated) read count matrix, \code{alpha}, a vector of
mutation-specific gene activation rates, \code{beta}, a vector
of mutation-specific gene deactivation rates, \code{kappa}, a numeric used to
calculate the sequencing error rate, \code{tau}, a numeric used to calculate
the sequencing error rate, \code{Klist}, a range of subclones to try,
\code{nchains}, the number of MCMC chains, and \code{pburn}, the percentage of
burn-in to remove.
}
\description{
Produce a list of phylogenetic trees in a sample space of size
(number of chains) x (number of subclones). Sampling is performed using
Metropolis-within-Gibbs. Output is of class \code{get_trees}.
}
\examples{
#Load post-processed data for patient GBM10
data("GBM10_postproc")

# Run Canopy2 to get list of phylogenetic trees corresponding to all chains
# and all subclones
get.trees.out<-get_trees(Rs=GBM10_postproc@Rs, Rb=GBM10_postproc@Rb,
                         Xs=GBM10_postproc@Xs, Xb=GBM10_postproc@Xb,
                         alpha=GBM10_postproc@param.est$alpha,
                         beta=GBM10_postproc@param.est$beta, kappa=1,
                         tau=999, Klist=4:6, niter=10000, nchains=20, thin=20,
                         pburn=0.5, ncores=4, seed=8675309)

# Examine diagnostic plots
get_diagnostics(get.trees.out, project=NULL, outpath=NULL)

# Get best tree across all chains and subclones via DIC
best.tree.out<-get_best_tree(get.trees.out)

best.tree.out

}
