% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_within_Gibbs.R
\name{MH_within_Gibbs}
\alias{MH_within_Gibbs}
\title{Internal function}
\usage{
MH_within_Gibbs(
  chain,
  K,
  nchains,
  niter,
  thin,
  niter.thin,
  burn.len,
  Rs,
  Xs,
  Rb,
  Xb,
  S,
  N,
  alpha,
  beta,
  kappa,
  tau,
  seed = 8675309
)
}
\description{
Gibbs with nested Metropolis-Hastings. See Alg. 2 in the main text for
further details.
}
\keyword{internal}
