% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBetaPois.R
\name{rBetaPois}
\alias{rBetaPois}
\title{The Beta-Poisson Distribution}
\usage{
rBetaPois(n, alpha, beta, scale = 1)
}
\arguments{
\item{n}{the number of random values to return}

\item{alpha}{a positive value, the alpha shape parameter of the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing}

\item{beta}{a positive value, the beta shape parameter of the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing.}

\item{scale}{a positive value, the scaling parameter for the beta distribution
estimated by using \code{get_burstiness_bpsc},
\code{get_burstiness_scale}, or other method of the user's choosing}
}
\value{
a vector of the random variates from the beta-Poisson distribution
}
\description{
This function is used to generate random variants for the beta-Poisson
distribution. The beta distribution is modified to have a support of
[0, \code{scale}] instead of the typical [0, 1], and \code{alpha} and
\code{beta} represent the shape parameters.
}
\details{
Each of the beta-Poisson random variates, \eqn{r}, are generated as

\deqn{r \sim Pois(\lambda),}

where \eqn{\lambda \sim c*Beta(\alpha,\beta)} for \eqn{c=}\code{scale},
\eqn{\alpha}=\code{alpha}, and \eqn{\beta}=\code{beta}.
}
\examples{
set.seed(8675309)
bp.vec=rBetaPois(100,0.1,1.0,3000)
hist(bp.vec, prob=T)

}
