% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampPs.R
\name{sampPs}
\alias{sampPs}
\title{Internal function}
\usage{
sampPs(tree, Rb, Xb, Rs, Xs, alpha, beta, kappa, tau, ...)
}
\description{
Sample the single-cell to clone assignment matrix, Ps, using Metropolis
Hastings (an accept reject algorithm). See Alg. 1 in the main text for
further details.
}
\keyword{internal}
